// src/app/model/factura.ts
/**
 * Proyecto TDS Backend
 * Descripción
 *
 * OpenAPI spec version: 2.0
 * Contact: info@tecazuay.edu.ec
 *
 * NOTE: This class is auto generated by the swagger code generator program.
 * https://github.com/swagger-api/swagger-codegen.git
 * Do not edit the class manually.
 */
import { Cliente } from './cliente';
import { CuerpoFactura } from './cuerpoFactura';

export interface Factura { 
    cliente?: Cliente;
    cuerpoFactura?: Array<CuerpoFactura>;
    fecha?: Date;
    idFactura?: number;
    tipoFactura?: string;
}